requirements.txt
# Core Dependencies
python-json-logger>=2.0.7
tenacity==8.2.3
ratelimit>=2.2.1
pytest>=8.0.0
python-dotenv==1.0.0
schedule>=1.2.0
cryptography>=41.0.3
dataclasses-json>=0.6.3
pydantic>=2.5.3
fastapi>=0.109.0
uvicorn>=0.25.0
redis>=5.0.1
prometheus-client>=0.19.0
circuitbreaker>=1.4.0
jsonschema>=4.21.1

# AI and NLP
openai==0.28.0
nltk>=3.8.1
transformers>=4.36.0
scikit-learn>=1.3.2
spacy>=3.7.2

# Task Management
asana==3.2.1
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
trello>=0.9.7

# Communication
twilio>=8.10.0
sendgrid>=6.10.0
linkedin-api>=2.0.0
praw>=7.7.1
facebook-business>=17.0.2
instabot>=0.117.0
tiktok-uploader>=1.0.16
phonenumbers>=8.13.20
slack-sdk>=3.26.1
python-telegram-bot>=20.7

# Web and Networking
aiohttp>=3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
selenium>=4.15.2
webdriver_manager>=3.0.0
playwright==1.39.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.3
pandas==2.0.3
pytesseract>=0.3.10
python-xlib>=0.33
pyautogui>=0.9.54

# Testing and Development
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
flake8>=6.1.0
